+++ lib/yard-sorbet/handlers/abstract_dsl_handler.rb
@@ -3,7 +3,7 @@
 
 module YARDSorbet
   module Handlers
-    # Apllies an `@abstract` tag to `abstract!`/`interface!` modules (if not alerady present).
+    # Applies an `@abstract` tag to `abstract!`/`interface!` modules (if not alerady present).
     class AbstractDSLHandler < YARD::Handlers::Ruby::Base
       extend T::Sig