+++ sorbet/tapioca/require.rb
@@ -1,6 +1,6 @@
 # typed: strong
 # frozen_string_literal: true
 
-# Add your extra requires here (`bin/tapioca require` can be used to boostrap this list)
+# Add your extra requires here (`bin/tapioca require` can be used to bootstrap this list)
 require 'rspec/core/rake_task'
 require 'rubocop/rake_task'